[AlertSlice]: Adding alert selector and updating its usages
