+++ src/store/selectors/auth/AuthSelector.tsx
