+++ src/store/selectors/auth/index.ts
@@ -1,1 +1,1 @@
-export * from './Auth.Selector';
+export * from './AuthSelector';
