+++ src/store/selectors/image_modal/ImageModalSelector.tsx
