+++ src/store/selectors/image_modal/index.ts
@@ -1,1 +1,1 @@
-export * from './ImageModal.Selector';
+export * from './ImageModalSelector';
