+++ src/store/selectors/index.ts
@@ -1,3 +1,5 @@
 export * from './auth';
 export * from './image_modal';
 export * from './notification';
+// export * from './alert';
+// export * from './profile';
