+++ src/store/selectors/notification/index.ts
@@ -1,1 +1,1 @@
-export * from './notification.selector';
+export * from './NotificationSelector';
