+++ src/store/selectors/profile/index.ts
@@ -1,1 +1,1 @@
-export * from './Profile.Selector';
+export * from './ProfileSelector';
