+++ src/store/slice/alert/index.ts
@@ -1,0 +1,1 @@
+export { alertSlice } from './AlertSlice';
