--- src/store/slice/toast_alert/index.ts
@@ -1,3 +1,0 @@
-export { default as alertReducer } from './ToastAlertSlice';
-export type { ToastAlertData } from './ToastAlertSlice';
-export { showAlertWithTimeout } from './ToastAlertSlice';
